src/requirements.txt
python-markdown
pymdown-extensions
jinja2
black
